*** new results ***
Same results, now with updated code. The strange bump in the pdf has disappeared, so I'd say it's an improvement!

It would also be interesting to look at multiple quantities of interest, like the concentration evolving in time. How hard is it to adapt the code for this case? (Probably easier then for the QMC...)